Start new project for refactoring
Started new project to refactor the code.
Created an event system and changed the Input class to work with it, instead of storing all the button presses on the keyboard in an array. Still, have to add mouse inputs to the input class.
TODO - create a rendering system
TODO - handle window creation according to the selected graphics api
TODO - implement a camera class